app/requirements.txt
kafka-python
websockets
python-dotenv
hdfs
pyspark
mysql-connector-python